Also moved FilterTooLongStats
